+++ ngx-codetab/src/lib/services/tabs.service.ts
@@ -8,11 +8,19 @@ export class TabService {
    addTab(tab: Tab) {
       if (this.tabs.length === 0) {
          tab.selected = true;
+      } else {
+          tab.selected = false;
       }
       this.tabs.push(tab);
+      return tab.selected;
    }
 
    getTabs(): Array<Tab> {
       return this.tabs;
+   }
+
+   isTabSelected(tab) {
+       const selectedTab = this.tabs.find((t) => t === tab);
+       return selectedTab.selected;
    }
 }